No markRefs unless renderMarks is provided
We have an application that uses react-range under the hood, and we
noticed that a range input was taking 2GB of RAM on our machines. I did
some investigation and found that regardless of whether the marks
functionality was being used, refs were being created for each possible
value of the range.

We have some fairly huge ranges (we're using the input to scrub a video
with potential microsecond accuracy), and can imagine that other people
are affected by the previous behavior. This change should allow us to
continue using large input ranges without incurring a memory penalty.